+++ R/range.R
@@ -1,9 +1,26 @@
 range_heuristic <- function(df) {
     if (nrow(df) == 0) return(c(xmin = NA_real_, xmax = NA_real_, ymin = NA_real_, ymax = NA_real_))
-    xleft <- ifelse(grepl("tile", df$piece_side), df$x-1, df$x-0.5)
-    xright <- ifelse(grepl("tile", df$piece_side), df$x+1, df$x+0.5)
-    yleft <- ifelse(grepl("tile", df$piece_side), df$y-1, df$y-0.5)
-    yright <- ifelse(grepl("tile", df$piece_side), df$y+1, df$y+0.5)
+
+    is_tile <- grepl("tile", df$piece_side)
+    xleft <- ifelse(is_tile, df$x-1, df$x-0.5)
+    xright <- ifelse(is_tile, df$x+1, df$x+0.5)
+    yleft <- ifelse(is_tile, df$y-1, df$y-0.5)
+    yright <- ifelse(is_tile, df$y+1, df$y+0.5)
+
+    #### dominoes
+
+    is_board <- grepl("board", df$piece_side)
+    xleft <- ifelse(is_board, df$x-0.5*df$rank, xleft)
+    xright <- ifelse(is_board, df$x+0.5*df$rank, xright)
+    yleft <- ifelse(is_board, df$y-0.5*df$rank, yleft)
+    yright <- ifelse(is_board, df$y+0.5*df$rank, yright)
+
+    is_board2 <- is_board & grepl("2", df$cfg)
+    xleft <- ifelse(is_board2, df$x-df$rank, xleft)
+    xright <- ifelse(is_board2, df$x+df$rank, xright)
+    yleft <- ifelse(is_board2, df$y-df$rank, yleft)
+    yright <- ifelse(is_board2, df$y+df$rank, yright)
+
     list(xmin = min(xleft), xmax = max(xright), ymin = min(yleft), ymax = max(yright))
 }